Move isRunningInCE and isRunningInEE in their own helper
