+++ gitlab/resource_gitlab_project_push_rules_test.go
@@ -1,9 +1,7 @@
 package gitlab
 
 import (
-	"errors"
 	"fmt"
-	"strings"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/resource"
@@ -12,26 +10,6 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-sdk/helper/acctest"
 )
-
-func isRunningInEE() (bool, error) {
-	if conn, ok := testAccProvider.Meta().(*gitlab.Client); ok {
-		version, _, err := conn.Version.GetVersion()
-		if err != nil {
-			return false, err
-		}
-		if strings.Contains(version.String(), "-ee") {
-			return true, nil
-		}
-	} else {
-		return false, errors.New("Provider not initialized, unable to get GitLab connection")
-	}
-	return false, nil
-}
-
-func isRunningInCE() (bool, error) {
-	isEE, err := isRunningInEE()
-	return !isEE, err
-}
 
 func TestAccGitlabProjectPushRules_basic(t *testing.T) {
 	var pushRules gitlab.ProjectPushRules
